Refactor and fix issue when some entry is in edit mode and user tries to insert new field.
