+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/editors/types/listview/DataTypeListView.java
@@ -22,7 +22,6 @@ import java.util.Optional;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.event.Observes;
 import javax.inject.Inject;
 
 import com.google.gwt.event.dom.client.ClickEvent;
@@ -38,7 +37,6 @@ import org.jboss.errai.ui.shared.api.annotations.Templated;
 import org.kie.workbench.common.dmn.client.editors.common.RemoveHelper;
 import org.kie.workbench.common.dmn.client.editors.types.common.DataType;
 import org.kie.workbench.common.dmn.client.editors.types.common.ScrollHelper;
-import org.kie.workbench.common.dmn.client.editors.types.listview.common.DataTypeEditModeToggleEvent;
 import org.uberfire.client.views.pfly.selectpicker.ElementHelper;
 
 import static org.kie.workbench.common.dmn.client.editors.types.common.HiddenHelper.hide;
@@ -96,8 +94,6 @@ public class DataTypeListView implements DataTypeList.View {
     private final ScrollHelper scrollHelper;
 
     private DataTypeList presenter;
-
-    private DataTypeListItem currentEditingItem;
 
     @Inject
     public DataTypeListView(final HTMLDivElement listItems,
@@ -365,25 +361,6 @@ public class DataTypeListView implements DataTypeList.View {
     @Override
     public HTMLDivElement getListItems() {
         return listItems;
-    }
-
-    public void onDataTypeEditModeToggle(final @Observes DataTypeEditModeToggleEvent event) {
-        if (event.isEditModeEnabled()) {
-            if (getCurrentEditingItem() != null) {
-                getCurrentEditingItem().disableEditMode();
-            }
-            setCurrentEditingItem(event.getItem());
-        } else if (event.getItem() == getCurrentEditingItem()) {
-            setCurrentEditingItem(null);
-        }
-    }
-
-    DataTypeListItem getCurrentEditingItem() {
-        return currentEditingItem;
-    }
-
-    void setCurrentEditingItem(final DataTypeListItem currentEditingItem) {
-        this.currentEditingItem = currentEditingItem;
     }
 
     @Override
