+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/editors/types/listview/common/DataTypeEditModeToggleEvent.java
@@ -34,6 +34,6 @@ public class DataTypeEditModeToggleEvent implements UberFireEvent {
     }
 
     public DataTypeListItem getItem() {
-        return this.item;
+        return item;
     }
 }
